Add Governance Support to Subgraph

+++ packages/nouns-contracts/contracts/governance/GovernorNDelegator.sol
@@ -8,45 +8,52 @@
 
 pragma solidity ^0.8.4;
 
-import "./GovernorNInterfaces.sol";
+import './GovernorNInterfaces.sol';
 
 contract GovernorNDelegator is GovernorNDelegatorStorage, GovernorNEvents {
-	constructor(
-			address timelock_,
-			address nouns_,
-            address vetoer_,
-			address admin_,
-	        address implementation_,
-	        uint votingPeriod_,
-	        uint votingDelay_,
-            uint proposalThresholdBPS_,
-            uint quorumVotesBPS_) {
-
+    constructor(
+        address timelock_,
+        address nouns_,
+        address vetoer_,
+        address admin_,
+        address implementation_,
+        uint256 votingPeriod_,
+        uint256 votingDelay_,
+        uint256 proposalThresholdBPS_,
+        uint256 quorumVotesBPS_
+    ) {
         // Admin set to msg.sender for initialization
         admin = msg.sender;
 
-        delegateTo(implementation_, abi.encodeWithSignature("initialize(address,address,address,uint256,uint256,uint256,uint256)",
-                                                            timelock_,
-                                                            nouns_,
-                                                            vetoer_,
-                                                            votingPeriod_,
-                                                            votingDelay_,
-                                                            proposalThresholdBPS_,
-                                                            quorumVotesBPS_));
+        delegateTo(
+            implementation_,
+            abi.encodeWithSignature(
+                'initialize(address,address,address,uint256,uint256,uint256,uint256)',
+                timelock_,
+                nouns_,
+                vetoer_,
+                votingPeriod_,
+                votingDelay_,
+                proposalThresholdBPS_,
+                quorumVotesBPS_
+            )
+        );
 
         _setImplementation(implementation_);
 
-		admin = admin_;
-	}
+        admin = admin_;
+    }
 
-
-	/**
+    /**
      * @notice Called by the admin to update the implementation of the delegator
      * @param implementation_ The address of the new implementation for delegation
      */
     function _setImplementation(address implementation_) public {
-        require(msg.sender == admin, "GovernorNDelegator::_setImplementation: admin only");
-        require(implementation_ != address(0), "GovernorNDelegator::_setImplementation: invalid implementation address");
+        require(msg.sender == admin, 'GovernorNDelegator::_setImplementation: admin only');
+        require(
+            implementation_ != address(0),
+            'GovernorNDelegator::_setImplementation: invalid implementation address'
+        );
 
         address oldImplementation = implementation;
         implementation = implementation_;
@@ -69,7 +76,7 @@ contract GovernorNDelegator is GovernorNDelegatorStorage, GovernorNEvents {
         }
     }
 
-	/**
+    /**
      * @dev Delegates execution to an implementation contract.
      * It returns to the external caller whatever the implementation returns
      * or forwards reverts.
@@ -79,12 +86,16 @@ contract GovernorNDelegator is GovernorNDelegatorStorage, GovernorNEvents {
         (bool success, ) = implementation.delegatecall(msg.data);
 
         assembly {
-              let free_mem_ptr := mload(0x40)
-              returndatacopy(free_mem_ptr, 0, returndatasize())
+            let free_mem_ptr := mload(0x40)
+            returndatacopy(free_mem_ptr, 0, returndatasize())
 
-              switch success
-              case 0 { revert(free_mem_ptr, returndatasize()) }
-              default { return(free_mem_ptr, returndatasize()) }
+            switch success
+            case 0 {
+                revert(free_mem_ptr, returndatasize())
+            }
+            default {
+                return(free_mem_ptr, returndatasize())
+            }
         }
     }
