+++ packages/nouns-contracts/contracts/test/TimelockHarness.sol
@@ -2,17 +2,14 @@
 
 pragma solidity ^0.8.4;
 
-
-import "../governance/Timelock.sol";
+import '../governance/Timelock.sol';
 
 interface Administered {
-    function _acceptAdmin() external returns (uint);
+    function _acceptAdmin() external returns (uint256);
 }
 
 contract TimelockHarness is Timelock {
-    constructor(address admin_, uint delay_)
-        Timelock(admin_, delay_) {
-    }
+    constructor(address admin_, uint256 delay_) Timelock(admin_, delay_) {}
 
     function harnessSetPendingAdmin(address pendingAdmin_) public {
         pendingAdmin = pendingAdmin_;
@@ -24,7 +21,7 @@ contract TimelockHarness is Timelock {
 }
 
 contract TimelockTest is Timelock {
-    constructor(address admin_, uint delay_) Timelock(admin_, 2 days) {
+    constructor(address admin_, uint256 delay_) Timelock(admin_, 2 days) {
         delay = delay_;
     }
